requirements.txt
Flask==2.3.2
google-auth==2.23.2
google-auth-oauthlib==1.1.0
requests==2.31.0
pytz
